fix: end-to-end integration test fixes across all frontends
- IR: Leaf value widened from f32 to f64 to preserve LightGBM's native
  f64 leaf precision; backend tree functions now return f64 accordingly
- XGBoost: strip XGBoost 3.x bracket notation from base_score (e.g.
  "[0E0]"); apply logit when base_score is in probability space; remove
  explicit base_score=0.0 from generate.py (causes degenerate training)
- LightGBM: handle both string ("binary sigmoid:1") and array
  (["binary", "crossentropy"]) objective formats
- sklearn ONNX: switch generate.py scripts from HistGradientBoosting to
  GradientBoosting (skl2onnx 1.20 compatibility); fix numpy 2.x int64
  casting in generators.py
- Integration tests: detect model.json alongside .zip and .onnx; add
  per-test-case tolerance override; XGBoost regression uses 1e-4 to
  account for f32 leaf accumulation error over 50 trees

+++ assets/tests/common/generators.py
@@ -107,8 +107,8 @@ def add_categorical_features(
 
         categorical_arrays.append(cat_feature)
 
-        # Add effect to target
-        # For each level, add/subtract from target
+        # Add effect to target (cast to float to support both regression and classification)
+        y = y.astype(float)
         for j, level in enumerate(levels):
             mask = cat_feature == level
             if j == 0:
